+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/Polls/PollResultsView.swift
@@ -52,8 +52,7 @@ struct PollResultsView<Factory: ViewFactory>: View {
             }
         }
         .background(Color(colors.background).ignoresSafeArea())
-        .navigationBarBackground()
-        .toolbar {
+        .toolbarThemed {
             ToolbarItem(placement: .principal) {
                 Text(L10n.Message.Polls.Toolbar.resultsTitle)
                     .bold()
@@ -66,7 +65,6 @@ struct PollResultsView<Factory: ViewFactory>: View {
                 } label: {
                     Image(systemName: "xmark")
                 }
-                .accentColor(colors.navigationBarTintColor)
             }
         }
         .navigationBarTitleDisplayMode(.inline)